+++ internal/updater/manager.go
@@ -32,7 +32,6 @@ const (
 
 	httpTimeout      = 60 * time.Second
 	downloadTimeout  = 20 * time.Minute
-	restartDelay     = 1200 * time.Millisecond
 	maxVersionParts  = 4
 	maxBackoffDelay  = 30 * time.Minute
 	defaultUserAgent = "AnimateAutoTool-Updater/1.0"